+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/controller/admin/project/vo/ProjectCostStatsRespVO.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Schema(description = "管理后台 - 项目管理 Response VO")
@@ -11,10 +12,10 @@ import java.util.Map;
 @ToString(callSuper = true)
 public class ProjectCostStatsRespVO {
     @Schema(description = "合同应收")
-    private Long contractAmount;
+    private BigDecimal contractAmount;
 
     @Schema(description = "合同已收")
-    private Long contractReceivedAmount;
+    private BigDecimal contractReceivedAmount;
 
 
     @Schema(description = "物资成本")
